Move `get_empty_block` helper to break a circular import dependency
